[DPE-10073] feat: consumer lag metrics colelction
